Combined Startup services to Program.cs, created new entities

+++ Controllers/DeliveriesController.cs
@@ -11,6 +11,7 @@ using Microsoft.AspNetCore.Authorization;
 
 namespace CourierWebApp.Controllers
 {
+    [Authorize]
     public class DeliveriesController : Controller
     {
         private readonly ApplicationDbContext _context;
@@ -21,29 +22,33 @@ namespace CourierWebApp.Controllers
         }
 
         // GET: Deliveries
-        [Authorize]
         public async Task<IActionResult> Index()
         {
-            var applicationDbContext = _context.Delivery.Include(d => d.Unit);
+            var applicationDbContext = _context.Delivery.Include(d => d.Customer).Include(d => d.Item).Include(d => d.Unit);
             return View(await applicationDbContext.ToListAsync());
         }
 
         // GET Deliveries/ShowSearchForm
-        [Authorize]
-        public async Task<IActionResult> ShowSearchForm()
+        public IActionResult ShowSearchForm()
         {
             return View();
         }
 
         // POST Deliveries/ShowSearchResults
-        [Authorize]
-        public async Task<IActionResult> ShowSearchResults(String SearchPhrase)
+        public async Task<IActionResult> ShowSearchResults(string SearchPhrase)
         {
-            return View("Index", await _context.Delivery.Where( d => d.Name.Contains(SearchPhrase)).ToListAsync());
+            var deliveries = await _context.Delivery
+                .Include(d => d.Customer)
+                .Include(d => d.Unit)
+                .Include(d => d.DeliveryItems) // Include the join table data
+                .ThenInclude(di => di.Item) // Then include the Item data from the join table
+                .Where(d => d.DeliveryItems.Any(di => di.Item.Name.Contains(SearchPhrase)))
+                .ToListAsync();
+
+            return View("Index", deliveries);
         }
 
         // GET: Deliveries/Details/5
-        [Authorize]
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
@@ -52,6 +57,8 @@ namespace CourierWebApp.Controllers
             }
 
             var delivery = await _context.Delivery
+                .Include(d => d.Customer)
+                .Include(d => d.Item)
                 .Include(d => d.Unit)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (delivery == null)
@@ -63,9 +70,10 @@ namespace CourierWebApp.Controllers
         }
 
         // GET: Deliveries/Create
-        [Authorize]
         public IActionResult Create()
         {
+            ViewData["CustomerId"] = new SelectList(_context.Customer, "CustomerId", "CustomerEmail");
+            ViewData["ItemId"] = new SelectList(_context.Item, "ItemId", "Description");
             ViewData["UnitId"] = new SelectList(_context.Unit, "UnitId", "Unit");
             return View();
         }
@@ -73,23 +81,34 @@ namespace CourierWebApp.Controllers
         // POST: Deliveries/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Name,UnitId,Price,Courier,Date,Quantity")] Delivery delivery)
+        public async Task<IActionResult> Create([Bind("Id,CustomerId,ItemId,UnitId,Quantity,Price,Date")] Delivery delivery)
         {
             if (ModelState.IsValid)
             {
                 _context.Add(delivery);
                 await _context.SaveChangesAsync();
+
+                // Assuming ItemId is the ID of the selected item for this delivery
+                var deliveryItem = new DeliveryItem
+                {
+                    DeliveryId = delivery.Id,
+                    ItemId = delivery.ItemId
+                };
+
+                _context.DeliveryItem.Add(deliveryItem);
+                await _context.SaveChangesAsync();
+
                 return RedirectToAction(nameof(Index));
             }
+            ViewData["CustomerId"] = new SelectList(_context.Customer, "CustomerId", "CustomerEmail", delivery.CustomerId);
+            ViewData["ItemId"] = new SelectList(_context.Item, "ItemId", "Description", delivery.ItemId);
             ViewData["UnitId"] = new SelectList(_context.Unit, "UnitId", "Unit", delivery.UnitId);
             return View(delivery);
         }
 
         // GET: Deliveries/Edit/5
-        [Authorize]
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -102,6 +121,8 @@ namespace CourierWebApp.Controllers
             {
                 return NotFound();
             }
+            ViewData["CustomerId"] = new SelectList(_context.Customer, "CustomerId", "CustomerEmail", delivery.CustomerId);
+            ViewData["ItemId"] = new SelectList(_context.Item, "ItemId", "Description", delivery.ItemId);
             ViewData["UnitId"] = new SelectList(_context.Unit, "UnitId", "Unit", delivery.UnitId);
             return View(delivery);
         }
@@ -109,10 +130,9 @@ namespace CourierWebApp.Controllers
         // POST: Deliveries/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [Authorize]
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Name,UnitId,Price,Courier,Date,Quantity")] Delivery delivery)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,CustomerId,ItemId,UnitId,Quantity,Price,Date")] Delivery delivery)
         {
             if (id != delivery.Id)
             {
@@ -139,12 +159,13 @@ namespace CourierWebApp.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
+            ViewData["CustomerId"] = new SelectList(_context.Customer, "CustomerId", "CustomerEmail", delivery.CustomerId);
+            ViewData["ItemId"] = new SelectList(_context.Item, "ItemId", "Description", delivery.ItemId);
             ViewData["UnitId"] = new SelectList(_context.Unit, "UnitId", "Unit", delivery.UnitId);
             return View(delivery);
         }
 
         // GET: Deliveries/Delete/5
-        [Authorize]
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
@@ -153,6 +174,8 @@ namespace CourierWebApp.Controllers
             }
 
             var delivery = await _context.Delivery
+                .Include(d => d.Customer)
+                .Include(d => d.Item)
                 .Include(d => d.Unit)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (delivery == null)
@@ -164,14 +187,26 @@ namespace CourierWebApp.Controllers
         }
 
         // POST: Deliveries/Delete/5
-        [Authorize]
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var delivery = await _context.Delivery.FindAsync(id);
-            _context.Delivery.Remove(delivery);
-            await _context.SaveChangesAsync();
+            var delivery = await _context.Delivery
+                .Include(d => d.DeliveryItems) // Include the related DeliveryItems
+                .FirstOrDefaultAsync(d => d.Id == id);
+
+            if (delivery != null)
+            {
+                // Remove related DeliveryItem records first
+                _context.DeliveryItem.RemoveRange(delivery.DeliveryItems);
+
+                // Now remove the Delivery record
+                _context.Delivery.Remove(delivery);
+
+                // Save the changes to the database
+                await _context.SaveChangesAsync();
+            }
+
             return RedirectToAction(nameof(Index));
         }
